pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "certamen-core"
dynamic = ["version"]
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE"]
authors = [
  { name="Nikolay Eremeev", email="nikolay.eremeev@outlook.com" },
]
maintainers = [
  { name="Nikolay Eremeev", email="nikolay.eremeev@outlook.com" },
]
description = "Certamen Core: Tournament-based LLM evaluation framework for decision synthesis."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
keywords = ["llm", "ai", "tournament", "evaluation", "decision-making", "litellm"]
dependencies = [
    "litellm>=1.77.1,<1.84.0",
    "pyyaml~=6.0",
    "colorama~=0.4.6",
    "httpx~=0.28.1",
    "tiktoken>=0.12.0",
    "aiofiles>=24.1,<25.2",
    "scikit-learn~=1.8.0",
    "markdown-it-py~=4.0",
    "google-auth~=2.49",
    "pydantic~=2.12",
    "python-dotenv~=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<10",
    "pytest-asyncio~=1.2",
    "pytest-cov~=7.0",
    "pytest-xdist~=3.0",
    "hypothesis~=6.0",
    "coverage[toml]~=7.13",
    "black>=26.3.1,<27.0",
    "isort>=5.13,<9.0",
    "ruff>=0.15.10,<1.0",
    "mypy~=1.11",
    "types-colorama~=0.4",
    "types-PyYAML~=6.0",
    "types-aiofiles~=25.1",
    "pre-commit~=4.5",
    "mutmut~=3.5",
    "radon~=6.0",
    "import-linter~=2.0",
]


[project.urls]
"Homepage" = "https://github.com/nikolay-e/certamen-core"
"Bug Tracker" = "https://github.com/nikolay-e/certamen-core/issues"
"Repository" = "https://github.com/nikolay-e/certamen-core"
"Documentation" = "https://github.com/nikolay-e/certamen-core#readme"
"Changelog" = "https://github.com/nikolay-e/certamen-core/blob/main/CHANGELOG.md"

[project.scripts]
certamen = "certamen.interfaces.cli.main:run_from_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "certamen.__about__.__version__"}

[tool.setuptools.package-data]
certamen = ["*.yml", "*.yaml", "py.typed"]

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
norecursedirs = ["mutants", ".git", ".mypy_cache", ".pytest_cache", "__pycache__"]

# Evidence-based testing configuration
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "hypothesis: property-based tests using Hypothesis framework",
    "mutation: tests specifically designed for mutation testing",
]

[tool.coverage.run]
# Evidence: Branch coverage better predicts defect exposure than line coverage
branch = true
source = ["src/certamen"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
# Baseline coverage threshold for entire codebase (synchronized with Makefile)
# Core modules (tournament.py, base.py) have stricter 60% requirement in Makefile
fail_under = 40
skip_covered = true
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

# Hypothesis configuration (property-based testing)
# Evidence: Hypothesis finds edge cases that example-based tests miss
[tool.hypothesis]
max_examples = 100
derandomize = true
deadline = 2000  # 2 seconds per test case

[tool.ruff]
line-length = 79
target-version = "py310"
exclude = ["docs/", "site/", "build/", "dist/", "*.ipynb"]

[tool.ruff.lint]
# Comprehensive rule selection for quality and security
select = [
    "F",      # Pyflakes
    "E9",     # Runtime errors
    "W6",     # Deprecated features
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "I",      # isort
    "N",      # pep8-naming
    "PLE",    # Pylint errors
    "PLW",    # Pylint warnings
    "RUF",    # Ruff-specific
    "S",      # flake8-bandit (security)
    "UP",     # pyupgrade
]
ignore = [
    "E501",    # Line too long (handled by black)
    "RUF012",  # Mutable class attributes
    "S101",    # Assert in tests (expected)
    "S311",    # random.uniform() for jitter is fine (not crypto)
    "S603",    # subprocess call - we use it safely
    "S607",    # subprocess partial path - we use known commands
    "N811",    # Constant imported as non-constant (alias pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["certamen"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S108"]  # Allow assert and /tmp usage in tests

[tool.mypy]
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = false
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
files = ["src/"]
exclude = ["tests/", "^docs/", "^benchmarks/"]

[[tool.mypy.overrides]]
module = [
    "litellm", "litellm.*",
    "sklearn", "sklearn.*",
    "markdown_it", "markdown_it.*",
    "httpx", "httpx.*",
    "dotenv", "dotenv.*",
    "yaml", "yaml.*",
    "colorama", "colorama.*",
    "pydantic", "pydantic.*",
]
ignore_missing_imports = true
disable_error_code = ["attr-defined"]

[[tool.mypy.overrides]]
module = "certamen.infrastructure.config.models"
disable_error_code = ["misc"]

# Mutation testing configuration
# Evidence: Mutation score correlates with real fault detection
# Note: mutmut 3.x creates mutants/ dir, needs editable install
[tool.mutmut]
paths_to_mutate = [
    "src/certamen/domain/",
]
backup = false

# Import Linter - Enforce clean architecture layer boundaries
# Dependency direction: interfaces -> application -> infrastructure -> domain -> ports -> shared
[tool.importlinter]
root_package = "certamen"
include_external_packages = false

[[tool.importlinter.contracts]]
name = "Domain layer independence"
type = "forbidden"
source_modules = ["certamen.domain"]
forbidden_modules = [
    "certamen.infrastructure",
    "certamen.interfaces",
    "certamen.application",
]

[[tool.importlinter.contracts]]
name = "Ports layer independence"
type = "forbidden"
source_modules = ["certamen.ports"]
forbidden_modules = [
    "certamen.infrastructure",
    "certamen.interfaces",
    "certamen.application",
    "certamen.domain",
]

[[tool.importlinter.contracts]]
name = "Shared layer independence"
type = "forbidden"
source_modules = ["certamen.shared"]
forbidden_modules = [
    "certamen.infrastructure",
    "certamen.interfaces",
    "certamen.application",
    "certamen.domain",
    "certamen.ports",
]

[[tool.importlinter.contracts]]
name = "Clean Architecture layer boundaries"
type = "layers"
layers = [
    "certamen.interfaces",
    "certamen.application",
    "certamen.infrastructure",
    "certamen.domain",
    "certamen.ports",
    "certamen.shared",
]
